Move lru_cache into own module
